CBASP-23: Refactor initialization and setup/configuration
Motivation
----------
Changes providers so that they no longer depend on ClusterHelper, so that
you can use different clusters for session and caching. Also, simplifies
configuration by removing need for CouchbaseBucketFactory.

Modifications
-------------
Make CouchbaseBucketFactory obsolete. Remove dependency on ClusterHelper;
use Cluster instead and scope to provider. Change configuration so that a
the bucket name in the SessionState and OutputCaching sections must be the
same name as the bucket you wish to use in your
CouchbaseConfigurationSection. Also, the name of on the SessionState or
OutputCache must match the CouchbaseConfigSection, so that upon
initialization the provider can lookup the correct bucket configuration to
use.